WriteStartupLog() not implemented
