+++ Code/Framework/AzQtComponents/AzQtComponents/Application/AzQtTraceLogger.h
@@ -12,33 +12,22 @@
 
 #pragma once
 
-#include <AzFramework/Logging/LogFile.h>
 #include <AzCore/Debug/TraceMessageBus.h>
+#include <AzQtComponents/AzQtComponentsAPI.h>
 
 namespace AzQtComponents
 {
-    class AzQtTraceLogger : private AZ::Debug::TraceMessageBus::Handler
+    class AZ_QT_COMPONENTS_API AzQtTraceLogger 
     {
     public:
         AzQtTraceLogger();
         ~AzQtTraceLogger();
-
         void WriteStartupLog(char name[]);
 
     protected:
-        //////////////////////////////////////////////////////////////////////////
-        // AZ::Debug::TraceMessageBus::Handler overrides...
-        bool OnOutput(const char* window, const char* message) override;
-        //////////////////////////////////////////////////////////////////////////
-
-        struct LogMessage
-        {
-        public:
-            AZStd::string window;
-            AZStd::string message;
-        };
-        AZStd::vector<LogMessage> m_startupLogSink;
-        AZStd::unique_ptr<AzFramework::LogFile> m_logFile;
+        class Impl;
+        AZ_PUSH_DISABLE_DLL_EXPORT_MEMBER_WARNING
+        AZStd::unique_ptr<Impl> m_impl;
+        AZ_POP_DISABLE_DLL_EXPORT_MEMBER_WARNING
     };
-}
-
+} // namespace AzQtComponents
