+++ Code/Framework/AzQtComponents/AzQtComponents/azqtcomponents_files.cmake
@@ -9,6 +9,8 @@ set(FILES
     AzQtComponentsAPI.h
     Application/AzQtApplication.cpp
     Application/AzQtApplication.h
+	Application/AzQtTraceLogger.cpp
+	Application/AzQtTraceLogger.h
     Buses/DragAndDrop.h
     Buses/ShortcutDispatch.h
     DragAndDrop/MainWindowDragAndDrop.h
